# lib/api-zod/src/generated/types/sellerSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SellerSummary {
  id: number;
  name: string;
  whatsapp: string;
  location: string;
}